requirements.txt
PyYaml
PyNaCl
pyCrypto
websockets
